+++ src/main/java/techcourse/myblog/web/controller/UserController.java
@@ -4,20 +4,20 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindException;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.service.UserReadService;
 import techcourse.myblog.service.UserWriteService;
 import techcourse.myblog.service.dto.UserDto;
-import techcourse.myblog.support.argument.LoginUser;
-import techcourse.myblog.support.exception.LoginFailedException;
 import techcourse.myblog.support.validation.UserInfo;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.groups.Default;
-import java.util.Optional;
 
 @Controller
 public class UserController {
@@ -36,11 +36,16 @@ public class UserController {
         }
         return "signup";
     }
+    
+    @GetMapping("/users")
+    public String userList(Model model) {
+        model.addAttribute("users", userReadService.findAll());
+        return "user-list";
+    }
 
     @PostMapping("/users")
     public RedirectView createUser(@ModelAttribute("/signup") @Validated({Default.class, UserInfo.class}) UserDto userDto) {
         userWriteService.save(userDto);
-
         return new RedirectView("/login");
     }
 
@@ -56,14 +61,9 @@ public class UserController {
     @PostMapping("/login")
     public RedirectView login(HttpSession session,
                               @ModelAttribute("/login") @Validated(Default.class) UserDto userDto) {
-        Optional<User> loginUser = userReadService.findByEmailAndPassword(userDto);
-
-        if (loginUser.isPresent()) {
-            session.setAttribute("user", loginUser.get());
-            return new RedirectView("/");
-        }
-
-        throw new LoginFailedException();
+        User loginUser = userReadService.findByEmailAndPassword(userDto);
+        session.setAttribute("user", loginUser);
+        return new RedirectView("/");
     }
 
     @GetMapping("/logout")
@@ -71,41 +71,7 @@ public class UserController {
         session.invalidate();
         return new RedirectView("/");
     }
-
-    @GetMapping("/users")
-    public String userList(Model model) {
-        model.addAttribute("users", userReadService.findAll());
-        return "user-list";
-    }
-
-    @GetMapping("/mypage")
-    public String myPage() {
-        return "mypage";
-    }
-
-    @GetMapping("/mypage/edit")
-    public String createMyPageForm(Model model) {
-        if (!model.containsAttribute("userDto")) {
-            model.addAttribute("userDto", new UserDto("", "", ""));
-        }
-        return "mypage-edit";
-    }
-
-    @PutMapping("/mypage")
-    public RedirectView editUser(LoginUser loginUser,
-                                 @ModelAttribute("/mypage/edit") @Validated(UserInfo.class) UserDto userDto) {
-        userWriteService.update(loginUser.getUser(), userDto);
-
-        return new RedirectView("/mypage");
-    }
-
-    @DeleteMapping("/mypage")
-    public RedirectView removeUser(LoginUser loginUser) {
-        userWriteService.remove(loginUser.getUser());
-
-        return new RedirectView("/logout");
-    }
-
+    
     @ExceptionHandler(BindException.class)
     public RedirectView handleBindError(BindException e, RedirectAttributes redirectAttributes) {
         redirectAttributes.addFlashAttribute("userDto", new UserDto("", "", ""));
