+++ src/test/java/techcourse/myblog/service/UserReadServiceTests.java
@@ -1,28 +1,38 @@
 package techcourse.myblog.service;
 
 import org.junit.jupiter.api.Test;
-import techcourse.myblog.service.dto.UserDto;
 import techcourse.myblog.service.common.UserCommonServiceTests;
+import techcourse.myblog.service.dto.UserDto;
+import techcourse.myblog.support.exception.LoginFailedException;
 
 import java.util.Arrays;
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
 
 public class UserReadServiceTests extends UserCommonServiceTests {
     @Test
-    public void findByEmailAndPassword_test() {
+    void 이메일_패스워드_일치() {
         given(userRepository.findByEmailAndPassword(user.getEmail(), user.getPassword()))
                 .willReturn(Optional.of(user));
 
-        assertTrue(userReadService.findByEmailAndPassword(new UserDto("", user.getEmail(), "Passw0rd!")).isPresent());
-        assertFalse(userReadService.findByEmailAndPassword(new UserDto("", "e@mail.com", "Passw0rd!")).isPresent());
+        assertEquals(userReadService.findByEmailAndPassword(new UserDto("", user.getEmail(), user.getPassword())), user);
     }
-
+    
     @Test
-    public void findAll_test() {
+    void 이메일_패스워드_일치_실패() {
+        given(userRepository.findByEmailAndPassword(user.getEmail(), user.getPassword()))
+                .willReturn(Optional.of(user));
+        
+        assertThrows(LoginFailedException.class, () -> {
+            userReadService.findByEmailAndPassword(new UserDto("", "e@mail.com", "Passw0rd!"));
+        });
+    }
+    
+    @Test
+    void 유저_리스트_성공() {
         given(userRepository.findAll()).willReturn(Arrays.asList(user));
 
         userReadService.findAll().forEach(foundUser ->
